Struct container1_beta1::ProjectMethodsBuilder

pub struct ProjectMethodsBuilder<'a, C, NC, A> where NC: 'a, C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on project resources. It is not used directly, but through the Container hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate container1_beta1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use container1_beta1::Container;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Container::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *MethodBuilders*
// like `clusters_list(...)`, `operations_list(...)`, `zones_clusters_create(...)`, `zones_clusters_delete(...)`, `zones_clusters_get(...)`, `zones_clusters_list(...)`, `zones_operations_get(...)` and `zones_operations_list(...)`
// to build up your call.
let rb = hub.projects();

Methods

impl<'a, C, NC, A> ProjectMethodsBuilder<'a, C, NC, A>

fn zones_clusters_get(&self, project_id: &str, zone_id: &str, cluster_id: &str) -> ProjectZoneClusterGetMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Gets a specific cluster.

fn operations_list(&self, project_id: &str) -> ProjectOperationListMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Lists all operations in a project, across all zones.

fn zones_clusters_delete(&self, project_id: &str, zone_id: &str, cluster_id: &str) -> ProjectZoneClusterDeleteMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Deletes the cluster, including the Kubernetes master and all worker nodes.

Firewalls and routes that were configured at cluster creation are also deleted.

fn clusters_list(&self, project_id: &str) -> ProjectClusterListMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Lists all clusters owned by a project across all zones.

fn zones_operations_get(&self, project_id: &str, zone_id: &str, operation_id: &str) -> ProjectZoneOperationGetMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Gets the specified operation.

fn zones_operations_list(&self, project_id: &str, zone_id: &str) -> ProjectZoneOperationListMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Lists all operations in a project in a specific zone.

fn zones_clusters_list(&self, project_id: &str, zone_id: &str) -> ProjectZoneClusterListMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Lists all clusters owned by a project in the specified zone.

fn zones_clusters_create(&self, request: &CreateClusterRequest, project_id: &str, zone_id: &str) -> ProjectZoneClusterCreateMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Creates a cluster, consisting of the specified number and type of Google Compute Engine instances, plus a Kubernetes master instance.

The cluster is created in the project's default network.

A firewall is added that allows traffic into port 443 on the master, which enables HTTPS. A firewall and a route is added for each node to allow the containers on that node to communicate with all other instances in the cluster.

Finally, a route named k8s-iproute-10-xx-0-0 is created to track that the cluster's 10.xx.0.0/16 CIDR has been assigned.